+++ include/picongpu/fields/FieldE.hpp
@@ -82,6 +82,8 @@ namespace picongpu
         DataBoxType getHostDataBox();
 
         GridBuffer<ValueType,simDim>& getGridBuffer();
+        GridBuffer<ValueType, simDim> &getGridBuffer2();
+        void sync( );
 
         GridLayout<simDim> getGridLayout();
 
@@ -97,6 +99,7 @@ namespace picongpu
 
 
         GridBuffer<ValueType,simDim> *fieldE;
+        GridBuffer<ValueType,simDim> *fieldE2;
     };