+++ include/picongpu/fields/MaxwellSolver/DirSplitting/DirSplitting.hpp
@@ -34,6 +34,8 @@
 #include <pmacc/math/vector/TwistComponents.hpp>
 #include <pmacc/math/vector/compile-time/TwistComponents.hpp>
 #include <pmacc/dataManagement/DataConnector.hpp>
+#include <pmacc/cuSTL/cursor/navigator/compile-time/TwistAxesNavigator.hpp>
+#include <pmacc/cuSTL/cursor/accessor/TwistAxesAccessor.hpp>
 
 
 namespace picongpu
@@ -81,29 +83,81 @@ namespace dirSplitting
                           (sizeof(T_Dummy) != 0));
 #endif
     };
+
+    namespace result_of
+    {
+        template<typename T_SpaceTwist, typename T_ComponentTwist, typename T_Cursor>
+        struct twistVectorForDirSplitting
+        {
+            typedef pmacc::cursor::Cursor<pmacc::cursor::TwistAxesAccessor<T_Cursor, T_ComponentTwist>,
+                           pmacc::cursor::CT::TwistAxesNavigator<T_SpaceTwist>,
+                           T_Cursor> type;
+        };
+    } // namespace result_of
+
+    template<typename T_SpaceTwist, typename T_ComponentTwist, typename T_Cursor>
+    HDINLINE
+    typename result_of::twistVectorForDirSplitting<T_SpaceTwist, T_ComponentTwist, T_Cursor>::type
+    twistVectorForDirSplitting(const T_Cursor& cursor)
+    {
+        return typename result_of::twistVectorForDirSplitting<T_SpaceTwist, T_ComponentTwist, T_Cursor>::type(
+            pmacc::cursor::TwistAxesAccessor<T_Cursor, T_ComponentTwist>(),
+            pmacc::cursor::CT::TwistAxesNavigator<T_SpaceTwist>(),
+            cursor);
+    }
+
 } // namespace dirSplitting
 
     template< typename T_CurrentInterpolation >
     class DirSplitting: private dirSplitting::ConditionCheck< DirSplitting< T_CurrentInterpolation > >
     {
     private:
-        template<typename OrientationTwist,typename CursorE, typename CursorB, typename GridSize>
-        void propagate(CursorE cursorE, CursorB cursorB, GridSize gridSize) const
+        template<uint32_t pass, typename SpaceTwist, typename OrientationTwist,typename JSpaceTwist,typename CursorE, typename CursorB, typename CursorJ, typename GridSize>
+        void propagate(CursorE cursorE, CursorB cursorB,CursorJ cursorJ,CursorE old_cursorE, CursorB old_cursorB, GridSize gridSize) const
         {
             using namespace cursor::tools;
             using namespace pmacc::math;
 
-            auto gridSizeTwisted = twistComponents<OrientationTwist>(gridSize);
+            using SpaceTwistSimDim = typename CT::shrinkTo<SpaceTwist,simDim>::type;
+            auto gridSizeTwisted = twistComponents<
+                SpaceTwistSimDim
+            >(gridSize);
 
             /* twist components of the supercell */
-            using BlockDim = typename CT::TwistComponents<SuperCellSize, OrientationTwist>::type;
+            typedef typename pmacc::math::CT::make_Vector<simDim,boost::mpl::integral_c<int,0> >::type ZeroVector;
+
+            typedef typename pmacc::math::CT::AssignIfInRange<
+                ZeroVector,
+                boost::mpl::integral_c<int,0>,
+                typename pmacc::math::CT::At<SuperCellSize,typename SpaceTwist::x>::type
+            >::type VectorWith_X;
+
+            typedef typename pmacc::math::CT::AssignIfInRange<
+                VectorWith_X,
+                boost::mpl::integral_c<int,1>,
+                typename pmacc::math::CT::At<SuperCellSize,typename SpaceTwist::y>::type
+            >::type VectorWith_XY;
+
+            typedef typename pmacc::math::CT::AssignIfInRange<
+                VectorWith_XY,
+                boost::mpl::integral_c<int,2>,
+                typename pmacc::math::CT::At<SuperCellSize,typename SpaceTwist::z>::type
+            >::type BlockDim;
+
+
+            pmacc::math::Size_t<simDim> zoneSize(gridSizeTwisted);
+            zoneSize.x()=BlockDim::x::value;
 
             algorithm::kernel::ForeachBlock<BlockDim> foreach;
-            foreach(zone::SphericZone<3>(pmacc::math::Size_t<3>(BlockDim::x::value, gridSizeTwisted.y(), gridSizeTwisted.z())),
-                    cursor::make_NestedCursor(twistVectorFieldAxes<OrientationTwist>(cursorE)),
-                    cursor::make_NestedCursor(twistVectorFieldAxes<OrientationTwist>(cursorB)),
-                    DirSplittingKernel<BlockDim>((int)gridSizeTwisted.x()));
+            foreach(zone::SphericZone<simDim>(zoneSize),
+                    cursor::make_NestedCursor(dirSplitting::twistVectorForDirSplitting<SpaceTwistSimDim, OrientationTwist>(cursorE)),
+                    cursor::make_NestedCursor(dirSplitting::twistVectorForDirSplitting<SpaceTwistSimDim, OrientationTwist>(cursorB)),
+                    cursor::make_NestedCursor(dirSplitting::twistVectorForDirSplitting<SpaceTwistSimDim, OrientationTwist>(cursorJ)),
+                    cursor::make_NestedCursor(dirSplitting::twistVectorForDirSplitting<SpaceTwistSimDim, OrientationTwist>(old_cursorE)),
+                    cursor::make_NestedCursor(dirSplitting::twistVectorForDirSplitting<SpaceTwistSimDim, OrientationTwist>(old_cursorB)),
+                    DirSplittingKernel<pass,BlockDim,JSpaceTwist>((int)gridSizeTwisted.x()));
         }
+
     public:
 
         using CellType = cellType::Centered;
@@ -111,14 +165,25 @@ namespace dirSplitting
 
         DirSplitting(MappingDesc) {}
 
-        void update_beforeCurrent(uint32_t currentStep) const
+        void update_afterCurrent(uint32_t currentStep) const
         {
-            using GuardDim = SuperCellSize;
+            typedef SuperCellSize GuardDim;
 
             DataConnector &dc = Environment<>::get().DataConnector();
 
             auto fieldE = dc.get< FieldE >( FieldE::getName(), true );
             auto fieldB = dc.get< FieldB >( FieldB::getName(), true );
+            auto fieldJ = dc.get< FieldJ >( FieldJ::getName(), true );
+
+            using namespace cursor::tools;
+
+
+
+            if (laserProfiles::Selected::INIT_TIME > float_X(0.0))
+                LaserPhysics{}(currentStep);
+
+            __setTransactionEvent(fieldE->asyncCommunication(__getTransactionEvent()));
+            __setTransactionEvent(fieldB->asyncCommunication(__getTransactionEvent()));
 
             auto fieldE_coreBorder =
                 fieldE->getGridBuffer().getDeviceBuffer().
@@ -129,60 +194,97 @@ namespace dirSplitting
                 cartBuffer().view(GuardDim().toRT(),
                                   -GuardDim().toRT());
 
-            using namespace cursor::tools;
-            using namespace pmacc::math;
+            auto old_fieldE_coreBorder =
+                fieldE->getGridBuffer2().getDeviceBuffer().
+                       cartBuffer().view(GuardDim().toRT(),
+                                         -GuardDim().toRT());
+            auto old_fieldB_coreBorder =
+                fieldB->getGridBuffer2().getDeviceBuffer().
+                cartBuffer().view(GuardDim().toRT(),
+                                  -GuardDim().toRT());
 
-            pmacc::math::Size_t<3> gridSize = fieldE_coreBorder.size();
+            auto fieldJ_coreBorder =
+                fieldJ->getGridBuffer().getDeviceBuffer().
+                cartBuffer().view(GuardDim().toRT(),
+                                  -GuardDim().toRT());
 
+            typedef pmacc::math::CT::Int<0,1,2> Orientation_XJ;
+            typedef pmacc::math::CT::Int<0,1,2> Space_XJ;
 
-            using Orientation_X = pmacc::math::CT::Int<0,1,2>;
-            propagate<Orientation_X>(
+            pmacc::math::Size_t<simDim> gridSize = fieldE_coreBorder.size();
+
+            typedef pmacc::math::CT::Int<0,1,2> Orientation_X;
+            typedef pmacc::math::CT::Int<0,1,2> Space_X;
+            typedef pmacc::math::CT::Int<0,2,1> JDir_X;
+            propagate<1, Space_X,Orientation_X,JDir_X>(
                       fieldE_coreBorder.origin(),
                       fieldB_coreBorder.origin(),
+                      fieldJ_coreBorder.origin(),
+                      old_fieldE_coreBorder.origin(),
+                      old_fieldB_coreBorder.origin(),
                       gridSize);
 
             __setTransactionEvent(fieldE->asyncCommunication(__getTransactionEvent()));
             __setTransactionEvent(fieldB->asyncCommunication(__getTransactionEvent()));
 
-            using Orientation_Y = pmacc::math::CT::Int<1,2,0>;
-            propagate<Orientation_Y>(
+            typedef pmacc::math::CT::Int<1,2,0> Orientation_Y;
+            typedef pmacc::math::CT::Int<1,0,2> Space_Y;
+            typedef pmacc::math::CT::Int<0,1,2> JDir_Y;
+            propagate<1, Space_Y,Orientation_Y,JDir_Y>(
                       fieldE_coreBorder.origin(),
                       fieldB_coreBorder.origin(),
+                      fieldJ_coreBorder.origin(),
+                      old_fieldE_coreBorder.origin(),
+                      old_fieldB_coreBorder.origin(),
                       gridSize);
 
             __setTransactionEvent(fieldE->asyncCommunication(__getTransactionEvent()));
             __setTransactionEvent(fieldB->asyncCommunication(__getTransactionEvent()));
 
-            using Orientation_Z = pmacc::math::CT::Int<2,0,1>;
-            propagate<Orientation_Z>(
+
+    #if (SIMDIM==DIM3)
+            //! \todo: currently 3D: check this code if someone enable 3D
+            typedef pmacc::math::CT::Int<2,0,1> Orientation_Z;
+            typedef pmacc::math::CT::Int<2,0,1> Space_Z;
+            typedef pmacc::math::CT::Int<0,2,1> JDir_Z;
+            propagate<0,Space_Z,Orientation_Z,JDir_Z>(
                       fieldE_coreBorder.origin(),
                       fieldB_coreBorder.origin(),
+                      fieldJ_coreBorder.origin(),
+                      old_fieldE_coreBorder.origin(),
+                      old_fieldB_coreBorder.origin(),
                       gridSize);
 
-            if (laserProfiles::Selected::INIT_TIME > float_X(0.0))
-                LaserPhysics{}(currentStep);
+            __setTransactionEvent(fieldE->asyncCommunication(__getTransactionEvent()));
+            __setTransactionEvent(fieldB->asyncCommunication(__getTransactionEvent()));
+    #endif
+
+            propagate<1, Space_Y,Orientation_Y,JDir_Y>(
+                    fieldE_coreBorder.origin(),
+                    fieldB_coreBorder.origin(),
+                    fieldJ_coreBorder.origin(),
+                    old_fieldE_coreBorder.origin(),
+                    old_fieldB_coreBorder.origin(),
+                    gridSize);
 
             __setTransactionEvent(fieldE->asyncCommunication(__getTransactionEvent()));
             __setTransactionEvent(fieldB->asyncCommunication(__getTransactionEvent()));
 
-            dc.releaseData( FieldE::getName() );
-            dc.releaseData( FieldB::getName() );
+            propagate<1, Space_X,Orientation_X,JDir_X>(
+                    fieldE_coreBorder.origin(),
+                    fieldB_coreBorder.origin(),
+                    fieldJ_coreBorder.origin(),
+                    old_fieldE_coreBorder.origin(),
+                    old_fieldB_coreBorder.origin(),
+                    gridSize);
+
+            __setTransactionEvent(fieldE->asyncCommunication(__getTransactionEvent()));
+            __setTransactionEvent(fieldB->asyncCommunication(__getTransactionEvent()));
         }
 
-        void update_afterCurrent(uint32_t) const
+        void update_beforeCurrent(uint32_t) const
         {
-            DataConnector &dc = Environment<>::get().DataConnector();
 
-            auto fieldE = dc.get< FieldE >( FieldE::getName(), true );
-            auto fieldB = dc.get< FieldB >( FieldB::getName(), true );
-
-            EventTask eRfieldE = fieldE->asyncCommunication(__getTransactionEvent());
-            EventTask eRfieldB = fieldB->asyncCommunication(__getTransactionEvent());
-            __setTransactionEvent(eRfieldE);
-            __setTransactionEvent(eRfieldB);
-
-            dc.releaseData( FieldE::getName() );
-            dc.releaseData( FieldB::getName() );
         }
 
         static pmacc::traits::StringProperty getStringProperties()
@@ -191,6 +293,7 @@ namespace dirSplitting
             return propList;
         }
     };
+
 
 } // namespace maxwellSolver
 } // namespace fields