+++ include/pmacc/cuSTL/algorithm/kernel/ForeachBlock.hpp
@@ -44,7 +44,7 @@ namespace kernel
 {
 
 #ifndef FOREACH_KERNEL_MAX_PARAMS
-#define FOREACH_KERNEL_MAX_PARAMS 4
+#define FOREACH_KERNEL_MAX_PARAMS 8
 #endif
 
 namespace detail